requirements.txt
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.2.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
blinker==1.4
chardet==3.0.4
Click==7.0
colorama==0.3.7
configparser==3.5.0b2
cryptography==2.6.1
dicttoxml==1.7.4
distorm3==3.4.1
django-markdown-deux==1.0.5
dnspython==1.16.0
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.0.1
Flask==1.0.2
funcsigs==1.0.2
html5lib==1.0.1
ipaddress==1.0.17


jsonrpclib==0.1.7

keyring==17.1.1
keyrings.alt==3.1.1
lxml==4.3.2
markdown2==2.3.10
MarkupSafe==1.1.0
mechanize==0.2.5
more-itertools==4.2.0
numpy==1.16.2
olefile==0.46
openpyxl==2.4.9
pathlib2==2.3.3
Pillow==5.4.1
pluggy==0.8.0
py==1.7.0
pycairo==1.16.2
pycrypto==2.6.1
PyGObject==3.36.0
djangorestframework
markdown       
django-filter
django-cors-headers
mod_wsgi

pyOpenSSL==19.0.0
PyPDF2==1.26.0
pytest==3.10.1
pytz==2019.1
pyxdg==0.25
scandir==1.9.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
slowaes==0.1a1
soupsieve==1.8
unicodecsv==0.14.1

webencodings==0.5.1
Werkzeug==0.14.1
djangorestframework
markdown       
django-filter
django-cors-headers
mod_wsgi
python-environ